+++ inlang/source-code/message-bundle-component/src/stories/inlang-message-bundle.stories.ts
@@ -1,6 +1,7 @@
 import "./inlang-message-bundle.ts"
 import type { Meta, StoryObj } from "@storybook/web-components"
 import { html } from "lit"
+import { multipleMatcherBundle } from "@inlang/sdk/v2-mocks"
 
 const meta: Meta = {
 	component: "inlang-message-bundle",
@@ -10,5 +11,8 @@ const meta: Meta = {
 export default meta
 
 export const Props: StoryObj = {
-	render: () => html`<inlang-message-bundle></inlang-message-bundle> `,
+	render: () =>
+		html`<inlang-message-bundle
+			messageBundle=${JSON.stringify(multipleMatcherBundle)}
+		></inlang-message-bundle> `,
 }