+++ inlang/source-code/paraglide/paraglide-sveltekit/example/src/lib/ui/LanguageSwitcher.svelte
@@ -1,21 +1,21 @@
 <script>
-    import { availableLanguageTags, languageTag } from "$paraglide/runtime";
+    import { availableLanguageTags, languageTag } from "$lib/paraglide/runtime";
     import { i18n } from "$lib/i18n";
 	import { goto } from "$app/navigation"
     import { page } from "$app/stores";
 	import { get } from "svelte/store"
 
     /**
-     * @param { import("$paraglide/runtime").AvailableLanguageTag } newLanguage
+     * @param { import("$lib/paraglide/runtime").AvailableLanguageTag } newLanguage
      */
     function switchToLanguage(newLanguage) {
         const canonicalPath = i18n.route(get(page).url.pathname)
         const localisedPath = i18n.resolveRoute(canonicalPath, newLanguage)
-        goto(localisedPath)
+        goto( localisedPath)
     }
 
     /**
-     * @type {Record<import("$paraglide/runtime").AvailableLanguageTag, string>}
+     * @type {Record<import("$lib/paraglide/runtime").AvailableLanguageTag, string>}
      */
     const labels = {
         en: "🇬🇧 English",