+++ inlang/source-code/paraglide/paraglide-sveltekit/src/runtime/exclude.ts
@@ -3,18 +3,9 @@ import { normalize } from "./utils/path.js"
 export type ExcludeConfig = (string | RegExp)[]
 
 export function createExclude(excludeConfig: ExcludeConfig): (path: string) => boolean {
-	const checks: ((path: string) => boolean)[] = []
+	const checks: ((path: string) => boolean)[] = excludeConfig.map((exclude) =>
+		typeof exclude === "string" ? (path) => path === exclude : (path) => exclude.test(path)
+	)
 
-	for (const exclude of excludeConfig) {
-		if (typeof exclude === "string") {
-			checks.push((path) => path === exclude)
-		} else {
-			checks.push((path) => exclude.test(path))
-		}
-	}
-
-	return (path) => {
-		path = normalize(path)
-		return checks.some((check) => check(path))
-	}
+	return (path) => checks.some((check) => check(normalize(path)))
 }