+++ inlang/source-code/paraglide/paraglide-sveltekit/src/runtime/hooks/handle.ts
@@ -1,8 +1,13 @@
 import type { Handle } from "@sveltejs/kit"
-import { getPathInfo } from "../utils/get-path-info.js"
-import { base } from "$app/paths"
 import type { I18nConfig } from "../adapter.js"
+import { parseRoute } from "../utils/route.js"
+import { negotiateLanguagePreferences } from "@inlang/paraglide-js/internal/adapter-utils"
+import { base } from "$app/paths"
 import { dev } from "$app/environment"
+import type { RoutingStrategy } from "../strategy.js"
+import type { ParaglideLocals } from "../locals.js"
+
+const LANG_COOKIE_NAME = "paraglide:lang"
 
 /**
  * The default lang attribute string that's in SvelteKit's `src/app.html` file.
@@ -48,6 +53,7 @@ export type HandleOptions = {
 }
 
 export const createHandle = <T extends string>(
+	strategy: RoutingStrategy<T>,
 	i18n: I18nConfig<T>,
 	options: HandleOptions
 ): Handle => {
@@ -55,18 +61,49 @@ export const createHandle = <T extends string>(
 	const dirPlaceholder = options.textDirectionPlaceholder ?? "%paraglide.textDirection%"
 
 	return ({ resolve, event }) => {
-		const { lang } = getPathInfo(event.url.pathname, {
-			availableLanguageTags: i18n.runtime.availableLanguageTags,
-			defaultLanguageTag: i18n.defaultLanguageTag,
-			base,
-		})
+		const [localisedPath] = parseRoute(event.url.pathname as `/${string}`, base)
+		const langFromUrl = strategy.getLanguageFromLocalisedPath(localisedPath)
+
+		const langCookie = event.cookies.get(LANG_COOKIE_NAME)
+		const cookieLang = i18n.runtime.isAvailableLanguageTag(langCookie) ? langCookie : undefined
+
+		const negotiatedLanguagePreferences = negotiateLanguagePreferences(
+			event.request.headers.get("accept-language"),
+			i18n.runtime.availableLanguageTags
+		)
+		const negotiatedLanguage = negotiatedLanguagePreferences[0]
+
+		const lang = langFromUrl ?? cookieLang ?? negotiatedLanguage ?? i18n.defaultLanguageTag
+
+		if (lang !== langFromUrl) {
+			// redirect to the correct language
+			const localisedPathname = strategy.getLocalisedPath(localisedPath, lang)
+			return new Response(undefined, {
+				status: 302,
+				headers: {
+					Location: localisedPathname,
+				},
+			})
+		}
+
+		if (lang !== cookieLang) {
+			event.cookies.set(LANG_COOKIE_NAME, lang, {
+				maxAge: 31557600, //Math.round(60 * 60 * 24 * 365.25) = 1 year,
+				sameSite: "lax",
+				path: base || "/",
+			})
+		}
 
 		const textDirection = i18n.textDirection[lang as T] ?? "ltr"
 
-		event.locals.paraglide = {
+		const paraglideLocals: ParaglideLocals<T> = {
 			lang,
 			textDirection,
 		}
+
+		// @ts-expect-error
+		// The user needs to have the ParaglideLocals type in their app.d.ts file
+		event.locals.paraglide = paraglideLocals
 
 		return resolve(event, {
 			transformPageChunk({ html, done }) {
@@ -79,11 +116,15 @@ export const createHandle = <T extends string>(
 					html.includes(SVELTEKIT_DEFAULT_LANG_ATTRIBUTE)
 				) {
 					console.warn(
-						"It seems like you haven't replaced the `lang` attribute in your `src/app.html` file. \n" +
-							`Please replace the \`lang\` attribute with the correct placeholder:"\n\n` +
-							` - <html ${SVELTEKIT_DEFAULT_LANG_ATTRIBUTE}>\n` +
-							` + <html lang="${langPlaceholder}" dir="${dirPlaceholder}">` +
-							`\n\nThis message will not be shown in production.`
+						[
+							"It seems like you haven't replaced the `lang` attribute in your `src/app.html` file.",
+							"Please replace the `lang` attribute with the correct placeholder:",
+							"",
+							` - <html ${SVELTEKIT_DEFAULT_LANG_ATTRIBUTE}>`,
+							` + <html lang="${langPlaceholder}" dir="${dirPlaceholder}">`,
+							"",
+							"This message will not be shown in production.",
+						].join("\n")
 					)
 				}
 
@@ -93,14 +134,3 @@ export const createHandle = <T extends string>(
 	}
 }
 
-declare global {
-	// eslint-disable-next-line @typescript-eslint/no-namespace
-	namespace App {
-		interface Locals {
-			paraglide: {
-				lang: string
-				textDirection: "ltr" | "rtl"
-			}
-		}
-	}
-}