+++ inlang/source-code/badge/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @inlang/badge
 
+## 0.7.32
+
+### Patch Changes
+
+- Updated dependencies [64e30ee]
+  - @inlang/sdk@0.35.5
+
+## 0.7.31
+
+### Patch Changes
+
+- Updated dependencies [3b2e0a6]
+  - @lix-js/client@2.2.0
+  - @inlang/sdk@0.35.4
+
+## 0.7.30
+
+### Patch Changes
+
+- Updated dependencies [548bc9e]
+  - @lix-js/client@2.1.0
+  - @inlang/sdk@0.35.3
+
+## 0.7.29
+
+### Patch Changes
+
+- @inlang/sdk@0.35.2
+- @lix-js/client@2.0.1
+
+## 0.7.28
+
+### Patch Changes
+
+- Updated dependencies [00ad046]
+  - @lix-js/client@2.0.0
+  - @inlang/sdk@0.35.1
+
 ## 0.7.27
 
 ### Patch Changes