+++ inlang/source-code/paraglide/paraglide-rollup/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @inlang/paraglide-rollup
 
+## 1.0.69
+
+### Patch Changes
+
+- @inlang/paraglide-unplugin@1.6.13
+
+## 1.0.68
+
+### Patch Changes
+
+- @inlang/paraglide-unplugin@1.6.12
+
+## 1.0.67
+
+### Patch Changes
+
+- @inlang/paraglide-unplugin@1.6.11
+
+## 1.0.66
+
+### Patch Changes
+
+- @inlang/paraglide-unplugin@1.6.10
+
+## 1.0.65
+
+### Patch Changes
+
+- @inlang/paraglide-unplugin@1.6.9
+
 ## 1.0.64
 
 ### Patch Changes