+++ inlang/source-code/paraglide/paraglide-sveltekit/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @inlang/paraglide-sveltekit
 
+## 0.9.0
+
+### Minor Changes
+
+- 76e7f20: Add the `ParaglideLocals` interface for properly typing paraglide's locals. It should be added in `app.d.ts` under `Locals.paraglide`. The `init` CLI will do this automatically
+- 8db2224: Previously, if the default language was available on the root `/`, then it would still be available under `/[lang]`. This was never intended & no longer the case.
+
+  If you still want this behavior set the `prefixDefaultLanguage` to `"always"` in your `i18n` config.
+
+### Patch Changes
+
+- 6319391: Fix: Excluded routes now use the default language instead of incorrecly attempting to detect the language form the URL
+
+## 0.8.7
+
+### Patch Changes
+
+- @inlang/paraglide-js@1.9.1
+- @inlang/paraglide-vite@1.2.63
+
+## 0.8.6
+
+### Patch Changes
+
+- Updated dependencies [3b2e0a6]
+  - @lix-js/client@2.2.0
+  - @inlang/paraglide-js@1.9.1
+  - @inlang/paraglide-vite@1.2.62
+
+## 0.8.5
+
+### Patch Changes
+
+- Updated dependencies [548bc9e]
+  - @lix-js/client@2.1.0
+  - @inlang/paraglide-js@1.9.1
+  - @inlang/paraglide-vite@1.2.61
+
+## 0.8.4
+
+### Patch Changes
+
+- @inlang/paraglide-js@1.9.1
+- @lix-js/client@2.0.1
+- @inlang/paraglide-vite@1.2.60
+
+## 0.8.3
+
+### Patch Changes
+
+- Updated dependencies [00ad046]
+  - @lix-js/client@2.0.0
+  - @inlang/paraglide-js@1.9.1
+  - @inlang/paraglide-vite@1.2.59
+
 ## 0.8.2
 
 ### Patch Changes