+++ inlang/source-code/plugins/i18next/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @inlang/plugin-i18next
 
+## 4.14.4
+
+### Patch Changes
+
+- Updated dependencies [64e30ee]
+  - @inlang/sdk@0.35.5
+
+## 4.14.3
+
+### Patch Changes
+
+- @inlang/sdk@0.35.4
+
+## 4.14.2
+
+### Patch Changes
+
+- @inlang/sdk@0.35.3
+
+## 4.14.1
+
+### Patch Changes
+
+- @inlang/sdk@0.35.2
+
+## 4.14.0
+
+### Minor Changes
+
+- c64f346: increase batching to 50 for i18n plugin
+
+### Patch Changes
+
+- @inlang/sdk@0.35.1
+
 ## 4.13.41
 
 ### Patch Changes