+++ inlang/source-code/plugins/t-function-matcher/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @inlang/plugin-t-function-matcher
 
+## 0.6.26
+
+### Patch Changes
+
+- Updated dependencies [64e30ee]
+  - @inlang/sdk@0.35.5
+
+## 0.6.25
+
+### Patch Changes
+
+- @inlang/sdk@0.35.4
+
+## 0.6.24
+
+### Patch Changes
+
+- @inlang/sdk@0.35.3
+
+## 0.6.23
+
+### Patch Changes
+
+- @inlang/sdk@0.35.2
+- @inlang/plugin@2.4.13
+
+## 0.6.22
+
+### Patch Changes
+
+- @inlang/sdk@0.35.1
+- @inlang/plugin@2.4.12
+
 ## 0.6.21
 
 ### Patch Changes