+++ inlang/source-code/sdk/multi-project-test/CHANGELOG.md
@@ -4,6 +4,26 @@
 
 ### Patch Changes
 
+- Updated dependencies [2806698]
+  - @inlang/cli@2.18.0
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [3896fe6]
+  - @inlang/cli@2.17.0
+
+## null
+
+### Patch Changes
+
+- @inlang/cli@2.16.10
+
+## null
+
+### Patch Changes
+
 - @inlang/cli@2.16.9
 
 ## null