+++ inlang/source-code/settings-component/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @inlang/settings-component
 
+## 1.0.22
+
+### Patch Changes
+
+- Updated dependencies [64e30ee]
+  - @inlang/sdk@0.35.5
+
+## 1.0.21
+
+### Patch Changes
+
+- @inlang/sdk@0.35.4
+
+## 1.0.20
+
+### Patch Changes
+
+- @inlang/sdk@0.35.3
+
+## 1.0.19
+
+### Patch Changes
+
+- @inlang/sdk@0.35.2
+
+## 1.0.18
+
+### Patch Changes
+
+- @inlang/sdk@0.35.1
+
 ## 1.0.17
 
 ### Patch Changes