+++ inlang/source-code/telemetry/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @inlang/telemetry
 
+## 0.3.42
+
+### Patch Changes
+
+- Updated dependencies [64e30ee]
+  - @inlang/sdk@0.35.5
+
+## 0.3.41
+
+### Patch Changes
+
+- @inlang/sdk@0.35.4
+
+## 0.3.40
+
+### Patch Changes
+
+- @inlang/sdk@0.35.3
+
+## 0.3.39
+
+### Patch Changes
+
+- @inlang/sdk@0.35.2
+
+## 0.3.38
+
+### Patch Changes
+
+- @inlang/sdk@0.35.1
+
 ## 0.3.37
 
 ### Patch Changes